+++ HISTORY.rst
@@ -6,6 +6,10 @@ Release History
 0.7.2 (?)
 +++++++++
 - Suppress `KeyboardInterrupt` exceptions and exit with status code 130.
+- Added ability to skip downloading files based on checksum in `ia download`,
+  `Item.download()`, and `File.download()`.
+- `ia download` is now verbose by default. Output can be suppressed with the `--quiet`
+  flag.
 
 0.7.1 2014-08-25
 ++++++++++++++++